+++ src/mem/nvs/settings.hpp
@@ -65,6 +65,9 @@ public:
   out::track::CurrentLimit getCurrentLimitService() const;
   esp_err_t setCurrentLimitService(out::track::CurrentLimit value);
 
+  out::track::CurrentLimit getCurrentLimitUpdate() const;
+  esp_err_t setCurrentLimitUpdate(out::track::CurrentLimit value);
+
   uint8_t getLedDutyCycleBug() const;
   esp_err_t setLedDutyCycleBug(uint8_t value);